# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAAS 'Karamelos'",
        "description": "Sistema de punto de venta para local de 'Karamelos'",
        "contact": {
            "email": "samaelortiz2218@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=SAAS 'Karamelos'"
        }
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso, retorna usuario y cookie token"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "31234491f21a35d2de09d4fae7ee13c4",
                "responses": {
                    "200": {
                        "description": "Información del usuario"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Listar ventas",
                "operationId": "419ebb078c83262b05f7c639ae8b2c21",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar por fecha (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ventas paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Registrar una nueva venta",
                "operationId": "51f42c056032d143981f26f6c0888f24",
                "requestBody": {
                    "description": "Datos de la venta y productos",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clerk",
                                    "client",
                                    "payment_method",
                                    "cash",
                                    "card",
                                    "sale_details"
                                ],
                                "properties": {
                                    "clerk": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "client": {
                                        "type": "string",
                                        "example": "Pedro"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "card",
                                            "mixed"
                                        ],
                                        "example": "cash"
                                    },
                                    "cash": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "card": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "sale_details": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "grams"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "grams": {
                                                    "type": "number",
                                                    "example": 120
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venta creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/sales/{id}": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Ver detalles de una venta",
                "operationId": "b352670063e2f09edbfb74f39aef9b56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la venta con productos"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    }
                }
            }
        }
    },
...
}